try this

+++ tf/stylometry/WinRateStylo/train_wr_stylometry.py
@@ -284,7 +284,7 @@ class ScaffoldedViTAndWinRate(tf.keras.Model):
     else:
       pos_clocks = tf.zeros((tf.shape(pos)[0], 2), dtype=tf.int32)
 
-    combined = tf.concat([features1, features2, pos_clocks, pos], axis=-1)
+    combined = tf.concat([features1, features2, tf.cast(pos_clocks, tf.float32), pos], axis=-1)
     
     win_rate = self.wr_pred(combined, training=training)
